Fix Export function failing if no package has been generated before (first run)

+++ AssetStoreBatchMode/AssetStoreBatchMode.cs
@@ -330,7 +330,9 @@ public static class AssetStoreBatchMode
     }
     static void Export(Package package, string localRootPath, string toPath)
     {
-        File.Delete(toPath);
+
+        if(File.Exists(toPath))
+		    File.Delete(toPath);
 
         var guids = GetGUIDS(package, localRootPath);
         Debug.Log("[Asset Store Batch Mode] Number of assets to export: " + guids.Length);
